refactored get cart to be gin agnostic & update controller accordingly
