+++ pkg/calculator/calculator.go
@@ -4,20 +4,15 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"github.com/gin-gonic/gin"
 	"gorm.io/gorm"
-	"html/template"
-	db2 "interview/pkg/db"
 	"interview/pkg/entity"
-	"log"
 	"strconv"
-	"strings"
 )
 
 type Handler interface {
 	AddItemToCart(ctx context.Context, sessionID string, data CartItem) Response
 	DeleteCartItem(ctx context.Context, sessionID, cartItemID string) Response
-	GetCartData(c *gin.Context)
+	GetCartData(ctx context.Context, sessionID string) Response
 }
 
 type calculator struct {
@@ -137,39 +132,30 @@ func (cal *calculator) DeleteCartItem(ctx context.Context, sessionID, cartItemID
 	return Response{Code: 302, RedirectURL: "/"}
 }
 
-func (cal *calculator) GetCartData(c *gin.Context) {
-	data := map[string]interface{}{
-		"Error": c.Query("error"),
-		//"cartItems": cartItems,
+func (cal *calculator) GetCartData(ctx context.Context, sessionID string) Response {
+	if sessionID == "" {
+		return Response{
+			Code: 200,
+		}
 	}
-
-	cookie, err := c.Request.Cookie("ice_session_id")
-	if err == nil {
-		data["CartItems"] = getCartItemData(cookie.Value)
+	data := cal.getCartItemData(ctx, sessionID)
+	return Response{
+		Code: 200,
+		Data: data,
 	}
-
-	html, err := renderTemplate(data)
-	if err != nil {
-		log.Println(err)
-		c.AbortWithStatus(500)
-		return
-	}
-
-	c.Header("Content-Type", "text/html")
-	c.String(200, html)
 }
 
-func getCartItemData(sessionID string) (items []map[string]interface{}) {
-	db := db2.GetDatabase()
-	var cartEntity entity.CartEntity
-	result := db.Where(fmt.Sprintf("status = '%s' AND session_id = '%s'", entity.CartOpen, sessionID)).First(&cartEntity)
+func (cal *calculator) getCartItemData(ctx context.Context, sessionID string) (items []map[string]interface{}) {
+	db := cal.db
 
+	var cartEntity entity.CartEntity
+	result := db.WithContext(ctx).Where(fmt.Sprintf("status = '%s' AND session_id = '%s'", entity.CartOpen, sessionID)).First(&cartEntity)
 	if result.Error != nil {
 		return
 	}
 
 	var cartItems []entity.CartItem
-	result = db.Where(fmt.Sprintf("cart_id = %d", cartEntity.ID)).Find(&cartItems)
+	result = db.WithContext(ctx).Where(fmt.Sprintf("cart_id = %d", cartEntity.ID)).Find(&cartItems)
 	if result.Error != nil {
 		return
 	}
@@ -185,25 +171,4 @@ func getCartItemData(sessionID string) (items []map[string]interface{}) {
 		items = append(items, item)
 	}
 	return items
-}
-
-func renderTemplate(pageData interface{}) (string, error) {
-	// Read and parse the HTML template file
-	tmpl, err := template.ParseFiles("static/add_item_form.html")
-	if err != nil {
-		return "", fmt.Errorf("Error parsing template: %v ", err)
-	}
-
-	// Create a strings.Builder to store the rendered template
-	var renderedTemplate strings.Builder
-
-	err = tmpl.Execute(&renderedTemplate, pageData)
-	if err != nil {
-		return "", fmt.Errorf("Error parsing template: %v ", err)
-	}
-
-	// Convert the rendered template to a string
-	resultString := renderedTemplate.String()
-
-	return resultString, nil
 }
