+++ pkg/calculator/response.go
@@ -3,4 +3,5 @@ package calculator
 type Response struct {
 	Code        int
 	RedirectURL string
+	Data        interface{}
 }
